+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 IMPROVEMENTS:
 
-* resource/xray_watch: add functionality to apply `path_ant_filter` for the watched repositories. [#PR]()
+* resource/xray_watch: add functionality to apply `path_ant_filter` for `repository` and `all-repos` `watch_resource.type`.
+ [#PR](https://github.com/jfrog/terraform-provider-xray/pull/82)
 
 ## 1.7.0 (October 6, 2022). Tested on Artifactory 7.41.13 and Xray 3.57.6